requirements.txt
opencv-python-headless
h5py
matplotlib
numpy
scipy
tqdm
scikit-learn
pillow
ipykernel
tensorflow
streamlit